stackql_deploy::commands

Module common_args

Common Command Arguments

This module defines common command-line arguments that can be reused across different commands in the application.

Enums

FailureAction

Possible actions to take on failure

Functions

dry_run

Common argument for performing a dry run

env_file

Common argument for specifying an environment file

env_var

Common argument for setting additional environment variables

log_level

Common argument for setting the log level

on_failure

Common argument for specifying the action on failure

show_queries

Common argument for showing queries in the output logs

stack_dir

Common positional argument for the stack directory

stack_env

Common positional argument for the stack environment