stackql_deploy::utils::server

Function check_and_start_server

pub fn check_and_start_server()

Checks if the server is running and starts it if necessary.

This function checks if the server is local and needs to be started. If the server is not running, it attempts to start it with the specified host and port.

Arguments

host - A reference to the server host address.

port - The port number to check.

Behavior

If the server is already running locally, it will display a message indicating this.

If a remote server is specified, it will display a message indicating the remote connection.

If the server needs to be started, it will attempt to do so and indicate success or failure.